Remapping flash traits to use base traits
